requirements.txt
PyYAML == 6.0
websocket-client == 1.2.3
# TODO: Gtk
